requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==1.0.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
aiomysql==0.2.0

# Autenticación (para iteraciones futuras)
passlib[bcrypt]==1.7.4